Expand source code definition
Even if we only want people to have source code so we can inspect the
software (as opposed to make modified versions that they run), it is
tremendously useful to have everything they need to actually build a
new version.  The FSF has been working on this problem for a couple
decades and the language in their GNU General Public License is the
gold standard in that regard.

This patch expands the scope of what is "source code" to include build
scripts and everything else necessary to actually work with the
source.