Adjust the definition of "Token"
To allow DeFi projects to function under this safe harbour, it's necessary to redefine the restriction on Token financial interests. Specifically, we need to be sure that a DAO-type organization can function, where decentralized users pool effort and/or capital. Limiting the restriction to centralized companies, partnerships, and funds is the cleanest change that would allow this. The restriction would still prevent a token being ownership of a hedge fund, but would now allow use cases such as Ethereum's Proof of Stake.