Add to the disclosure requirements
It's necessary, but not sufficient, to post disclosures on a public website; the disclosures must be readily available to the public in the venues where the Initial Development team is talking about the network. Otherwise, I could post disclosures on www.thisisadisclosurethatyouwillnevereverfind.com, and be compliant with the rule, but ineffective at communicating the disclosure to the public.