Ensure outside counsel is external
Because the Exit Report can be filed by anyone, a strict reading of the text would allow an unconnected company to write the Exit Report, and have the Initial Development Team's internal counsel write the attestation (because to that unconnected company, the internal counsel is an outside counsel). This is obviously not what we want.