Expectation of Token and Network Maturity
For some Tokens, the network on which the Token functions is already at Network Maturity (an ERC-20 token on Ethereum). We need to ensure that there's no loophole here. Requiring both the Token and the network on which it functions to have reached Network Maturity prevents that loophole. Also, some Tokens function on multiple networks, so adding "or networks" here.

Note: This clause does prevent Tokens on centralized blockchains from taking this step, so it would not allow Token safe harbor on a chain such as Binance Chain, where it is not decentralized. I think this is a good thing, as it also ensures that any underlying blockchain is sufficiently decentralized, reducing the risk of a bad actor creating a decentralized project on a larger centralized network.