Lining up durations
The durations of the rule were contradictory, between (a)(1) - 3 years from the first token sale, and (e) - 3 years from filing the notice of reliance. Also, Section (h) allows filing as soon as practicable, meaning a prior Token, which has already had more time to reach Network Maturity, could gain extra time to reach Network Maturity that a new Token could not.

To that effect, make the duration the later of (a) three years after the date of first sale of tokens, or (b) three years after the rule is effective. The second ensures a level playing field for projects that previously sold tokens; their clock starts on a clearly-defined day. Apologies to the SEC employees who, three years after that date, have to sift through a couple of thousand exit reports.