Require Initial Dev Team to author the exit report
Until the exit report is filed, the Token is operating under safe harbour. While the token is operating under safe harbour, there is an Initial Development Team, and that team has legal accountability for the Token and any warranties made to the SEC or other entities. Therefore, the Initial Development Team must file the exit report