Widen language to include protocol and dApps
This fixes https://github.com/CommissionerPeirce/SafeHarbor2.0/issues/24. 

The core change is to explicitly state that this proposal includes protocols, and decentralized apps ("dApps") as well as blockchain networks. Such dApps (eg. Compound, or Uniswap) are built on top of blockchain networks, and constitute the majority of new projects in the space. So having clarity on this point is crucial. And since both dApps and blockchains often sell tokens and both aspire to full decentralization, it seems appropriate to include them.

Note: the original issue suggested other changes, but upon reading through it again, I think the one change to clarify the definition of "network" up top may be sufficient.