Update README.md
Proposed Edits relate to using the benefits of an open ledger system to provide notice directly on the blockchain, where it would be possible to have front-end developers sort the tokens for "reliance" on this safe harbor.   This notice would provide much needed innovation of putting these disclosures in a readable format in front of token buyers, markets, etc.

Additionally, the benefits of having an open ledger system with the Initial Development Team having public wallets would provide the entire ecosystem the tools to track and witness the holdings and Token sales of teams in real time.  A living and breathing disclosure system.  Seems important to leverage the promise and areas where the blockchain and provide "superior" disclosure than a rolling series of 6 month Edgar notice filings.